+++ crates/compiler/tests/fixtures/script/import_meta/url/node/output.library-browser-cjs.js
@@ -1,0 +1,2 @@
+//index.js:
+ console.log(new URL(__filename, "file:").href);