+++ js-plugins/solid/CHANGELOG.md
@@ -68,6 +68,16 @@
 - Updated dependencies [a82a5b2]
   - @farmfe/core@2.0.0-nightly-20241022014521
 
+## 2.10.0
+
+### Patch Changes
+
+- Updated dependencies [e63b163a]
+- Updated dependencies [e63b163a]
+- Updated dependencies [e63b163a]
+- Updated dependencies [e63b163a]
+  - @farmfe/core@1.5.0
+
 ## 2.9.0
 
 ### Patch Changes