+++ packages/runtime-plugin-import-meta/CHANGELOG.md
@@ -6,6 +6,12 @@
 
 - cab79e8: bump major version with runtime
 
+## 0.2.3
+
+### Patch Changes
+
+- 249e9a2b: bundle runtime packages
+
 ## 0.2.2
 
 ### Patch Changes